Enabled colour coding to differentiate between completed and uncompleted assignments.

+++ src/main/java/seedu/address/model/assignment/DeadlineComparator.java
@@ -14,9 +14,9 @@ public class DeadlineComparator implements Comparator<Assignment> {
             return -1;
         } else {
             if (a1.getDeadline().getDateTime().isEqual(a2.getDeadline().getDateTime())) {
-                if (Integer.parseInt(a1.getWorkload().estHours) == Integer.parseInt(a2.getWorkload().estHours)) {
+                if (Float.parseFloat(a1.getWorkload().estHours) == Float.parseFloat(a2.getWorkload().estHours)) {
                     return 0;
-                } else if (Integer.parseInt(a1.getWorkload().estHours) > Integer.parseInt(a2.getWorkload().estHours)) {
+                } else if (Float.parseFloat(a1.getWorkload().estHours) > Float.parseFloat(a2.getWorkload().estHours)) {
                     return 1;
                 } else {
                     return -1;
