+++ src/main/java/seedu/address/model/assignment/WorkloadComparator.java
@@ -13,7 +13,7 @@ public class WorkloadComparator implements Comparator<Assignment> {
         } else if (a2.getStatus().status.equals("Completed") && a1.getStatus().status.equals("Not completed")) {
             return -1;
         } else {
-            if (Integer.parseInt(a1.getWorkload().estHours) == Integer.parseInt(a2.getWorkload().estHours)) {
+            if (Float.parseFloat(a1.getWorkload().estHours) == Float.parseFloat(a2.getWorkload().estHours)) {
                 if (a1.getDeadline().getDateTime().isEqual(a2.getDeadline().getDateTime())) {
                     return 0;
                 } else if (a1.getDeadline().getDateTime().isAfter(a2.getDeadline().getDateTime())) {
@@ -21,7 +21,7 @@ public class WorkloadComparator implements Comparator<Assignment> {
                 } else {
                     return -1;
                 }
-            } else if (Integer.parseInt(a1.getWorkload().estHours) > Integer.parseInt(a2.getWorkload().estHours)) {
+            } else if (Float.parseFloat(a1.getWorkload().estHours) > Float.parseFloat(a2.getWorkload().estHours)) {
                 return -1;
             } else {
                 return 1;
